changed LED to BitIO, modified all files to match the LEDs
Now everything is named alike. No ..._LED anymore, everything ..._LEDS.
Yet, still issues with "undefined references" which are in my opinion
impossible. I reviewed the paths of the definitions like 10-times and
compared to Mr. Stygers solution and found no differences. It's now
almost like i copied everything. It's working with the robot, but not
with FRDM which leads to the assumption that there is an issue with the
Platform_local.h that i checked - couldn't find anything. I'm out of
ideas right know.